+++ tqsdk/__init__.py
@@ -4,8 +4,8 @@ __author__ = 'chengzhi'
 name = "tqsdk"
 
 from .__version__ import __version__
-from tqsdk.api import TqApi, TqAccount, TqChan, TqReplay
+from tqsdk.api import TqApi, TqAccount, TqChan
 from tqsdk.lib import TargetPosTask, InsertOrderUntilAllTradedTask, InsertOrderTask
-from tqsdk.backtest import TqBacktest
+from tqsdk.backtest import TqBacktest, TqReplay
 from tqsdk.sim import TqSim
 from tqsdk.exceptions import BacktestFinished
